requirements.txt
altgraph==0.17.2
asgiref==3.5.0
asn1crypto==1.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
distlib==0.3.4
Django==3.2.11
filelock==3.5.1
fonttools==4.29.1
future==0.18.2
idna==3.3
oscrypto==1.2.1
pbr==5.8.1
pefile==2021.9.3
Pillow==9.0.1
platformdirs==2.5.0
pycparser==2.21
python-barcode==0.13.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
tzdata==2021.5
tzlocal==4.1
uritools==4.0.0
urllib3==1.26.8
